+++ apps/bot/src/domains/crew/services/index.ts
@@ -1,0 +1,2 @@
+export { disembarkCharacter } from './disembark-character.js';
+export { embarkCharacter } from './embark-character.js';
